+++ CHANGELOG.md
@@ -31,6 +31,7 @@
   * Added the missing test database,
   * Reduced the threshold for the test results as the results on travis are lower than local,
   * Fixed a minor issue linked to mysql,
+  * Fifth release candidate,
   * ...,